+++ Tests/Fixtures/ObjcAnnotatedRetentionFixtures/testRetainsExtensionDeclarationsOnObjcMembersAnnotatedClass.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@objcMembers class FixtureClass217 {}
+
+extension FixtureClass217 {
+    func methodInExtension() {}
+}